Renaming from specialization to type.
