+++ src/example/exampleApi.ts
@@ -30,7 +30,7 @@ import {
 } from '@blinkk/selective-edit';
 import {AsideFieldConfig} from '../editor/field/aside';
 import {ExampleFieldConfig} from './field/exampleField';
-import {GrowDocumentConfig} from '../specialized/grow/field/document';
+import {GrowDocumentConfig} from '../projectType/grow/field/document';
 import {MediaFieldConfig} from '../editor/field/media';
 
 const MAX_RESPONSE_MS = 1200;
